Extend the default API response limit
At some point we'll need to revisit this. By default, Sails Blueprint APIs return populated models (builds associated with a site) ordered by ID. We want the most recent builds. For now, we expect that we'll get the whole list with a high defaultLimit and take the last few we need by filtering on the client, but this sends a lot of unecessary data.