Ensure Markdown document is formatted
We want to keep the document as similar as possible so that we don’t
create lots of git commit trashing on whitespace. To that end, we add a
new line character to the front matter if there isn’t one already, so
that the trailing delimiter for front matter is on it’s own line.