Use _navigation.json to determine project structure
Previously, all repositories were displayed as a list of files, which
the user could click through. Considering that the idea of file
browsing doesn’t always make the most sense, we have a new view to
display “Pages”. This view is used when the repository has a
`_navigation.json`, which describes which pages to show and in what
hierarchy.

In cases where there is no _navigation.json, our old file browser will
be used instead.