Use defaults from main config if present
Jekyll uses a config variable called defaults in _config.yml, so
we want to use the same thing. Instead of looking for a _defaults.yml
file, now searches for the _config.yml, retrieves it from Github (not
S3 like the other configs) and then uses the defaults setting for new
pages.

Also adds yamljs as a dependency to handle the yaml

 Changes to be committed:
	modified:   assets/app/models/Github.js
	modified:   package.json