Return all index content from initial load
It isn’t best practices for even the home page to need to be generated
and rendered by the client. For instance, that means that any device
without javascript not only couldn’t use Federalist but couldn’t learn
anything about it either. Puts all content that should be in the index
in the page to be returned on initial load