Show text from error when adding site
The poor UI that is reported in #311 is fixed and tested. The actual
error message was under a `raw` key in `responseJSON`. The error
handler method was slightly refactored so that the message generation
was easily testable.

Adds tests to the front end suite to test the message returning function