Reorganize views and templates to match new design
Renames most of the templates and moves them so that the template
directory structure is easier to reason with when using the designs and
views.

Changes the views and routes around to make sure that views
appropriately encapsulate logic given the new designs.

Adds styles to get the most basic version of the new design
implemented. Much CSS refactoring needed and many details are missing.

Makes sure view unit code continues to pass.