Add script to export all current sites
There is a business need to know which sites are currently on
Federalist and which users we should get in touch with regarding
downtime or maintenance.

I have added a script that is executable with `npm run export:sites`.
You can pass an optional path after `--` to have the outputted CSV
written to a location of your choice.

I don’t think we can get into production cloud.gov postgres instances
from the outside, so we’ll have to run this script on a throwaway
instance.