Use React instead of bel
Using regular React instead of a combo of bel and morphdom
so that we can use libraries that have been tested and are
widely used by the community.

As a result, lots of actions changed, the structure of the app
changed, and the way routing was handled changed as well.