Enable save when user changes site settings
We want to give the user some feedback about their changes,
so instead of implementing an alert (which we should do in the
future as well) I went ahead and just made it so that the Save
button is disabled unless the user has made changes to the text
fields.

When the user clicks save, I disable the button again.