Catch some runtime errors in front end
In the github api we need to make sure config is defined,
because if the file doesn't exist for that site then it
will be undefined.

With the pure function syntax, I don't think you use this.props
favoring just the props that is passed into the function.

Right after we add a new site, the site object in the store
might not have any builds, and so the key will be undefined.
To get the site listing (required since a router update is
dispatched on a successful site add) the component has to
handle the case where builds doesn't yet exist.