Add unit test coverage for all of the redux reducers.
Fixes #466.

Found what may be a defect when receiving assets for a site with a URL
that matches another site's URL. If a URL is really a URL, then it's
not a bug. If a URL is really a URI, well, we probably have a problem.

Found some confusing code-- the reducer seems to think the state could
have a root site childDirectoriesMap but there does not appear to be
any way we can set one of these. Recommend dropping any reference to
this root childDirectoriesMap until we find a need to implement it.