Add SITE_FILE_CONTENT_RECEIVED action
We want to fetch content from the Github API when
we’re editing a file. This response is slightly
different from the other responses as it returns
a single object with a base64 encoded content key.
This is the content that we’ll decode and then use
it to populate the codemirror and prosemirror editors.

We keep the content encoded in the state reducer for
ease of passing it around. All components that need
it can decode it on their own.

* Add an action for fetching a file’s content
* Add a handler to the site reducer for the
    SITE_FILE_CONTENT_RECEIVED action type
* Maintain site reducer test code coverage