Update editors with file's content from Github
The editor container will decode and split up the file's content,
and then push the correct content to the two different editors
via their props.

Update both editor components to not just use the initial props,
but to also update the editor content when new props are passed
by the container component.