Add new files instead of replacing the whole array
When a siteActionTypes.SITE_CONTENTS_RECEIVED is
handled, it shouldn’t remove any files that might
already be in the file array.

For example, if the user loads the site with a URL
for the editor, it could be the case that the file
object (representing the page being edited) is
already in the state. In that case, we should just
extend it and add the new files to the array.

Other things:
* Adjust tests to take account of the new behavior
* Use the path attribute of the file objects as a
  key/unique identifier
* Add some white space to make each case easier to
  read