<Editor/> gets updated with child editor content
The editor container component needs to be able
to get content out of child editors to do things
like saving the document. Using the onChange
prop, the editor component now updates state on
every change in either editor. This makes it
easy to compose the document as a string for
saving to Github.

Also added a few react lifecycle callbacks so
that the editor can ensure that the content
is fetched from Github and so that it responds
properly to new props.

Adds a basic commit message field.