Add throwRuntime() for debugging help
Some of the action handlers will throw
errors based on the reducer logic, and
then show that error message in the
alert component. On the one hand, this
is nice to see the error in a more
user facing way, but it also removes
any stack trace or mention of the
error in the console which makes actual
debugging more challenging.

I think the best way to handle this is
to make a generic catch handler that
will only trigger an http error action
for an actual http error and use it
across all action handlers.