Editor should be able to save a new page
The editor needs to handle the case of
loading existing content or creating a
new file. When the user wants to create
a new file, we’ll prompt for the file
name and use that as the path.

Also add some UI around the save message
form so that it is more obvious what the
user should use it to do.