Put a line above nested page children when using the navigation.json.
It looks unpleasant, but arguably slightly better than having the lack
of line in between rows before. Will look more closely at this later as
this is roughly the same quality we have on production. No need to hold
up the react-redux train as it reaches feature/function parity with the
existing backbone implementation.