Add test:client:watch task to npm
I wanted to run the test suite whenever I made a change in
test code or application code. I installed `watch` to do this
and am using the "test:client:watch" script on npm to trigger it.

I'm not sure if this is the best name for the script, and I'm
totally open to changing it.