Add factories for models
This commit adds "factories" which can be used to create a record for a model to be used in the tests.

The syntax to create a record looks like this:

```javascript
factory(Site, {
  repository: "custom repo name"
}).then(site => {
  // use the site here
})
```

The factory will assign reasonable defaults to required attributes.
You can pass for those values in the second argument.
Overrides can be either a value or a Promise for a value.