Add tests for add-site User API action
The site-api action adds a site that already has an existing repo to a user's account.
This is accessed at the following endpoint:

POST /v0/user/add-site

This commit adds tests for this endpoint, and cleans up a bug that was discovered while writing the tests.
The bug that was discovered enabled users to add sites for other users.