Add tests for login / logout routes
This commit adds tests for the basic login / logout tests.

These tests do not actually test the auth flow, since that is handled by the OAuth callback.
Instead they simply test that redirects work as expected and the session is handled appropriatly.
Tests for the auth flow will need to come later.