Add codemirror to as production dependency
Codemirror was not a production dep which caused problems deploying to staging.
This commit adds it to the package.json so it will be installed on production.

This commit also adds the node_modules to the cfignore since it seemed to work in GovCloud recently.
This commit also adds `uswds.min.css` to the gitignore so I don't have to keep manually removing it before comitting.