Fix webhook issue that was preventing site creation
This commit fixes an issue creating webhooks that was occuring when users attempted to add repository based sites.
The issues occured because `GitHub.setWebhook(site, user, done)` was being passed a user object instead of an integer for the user's id.

This commit fixes the issues and adds tests to verify this works as expected in the future.