Remediate ws vulnerability by disabling websockets entirely
This commit entirely disables the sockets and pubsub sails hooks so the DOS vulnerability in the `ws` modules is never reached.
We don't actually use websockets at this point so there is no real point in keeping this code around.

This commit also cleans up some old code paths that leaned on web sockets.