Remove staging environment configuration
This commit removes the staging environment configuration.
Currently the staging app is deployed with the `NODE_ENV` set to `production`.

The `APP_ENV` var is used to distinguish the environments in code, but the `NODE_ENV` is used to determine which env is loaded so that staging env never gets touched.