Initialize SQS object in SQS service
This commit changes the config files to use a traditional configuration object for SQS instead of an actual instance of an SQS client.
The SQS client is now initialized from the config within the SQS service class.

This will prevent issues that may arise if the SQS configuration is blank which could prevent the app from lifting since it won't be able to reach the SQS object in the config.