Replace /v0/user with /v0/me
/v0/user returned the current user.
This commit ammends that API response so that it is now found at /v0/me and is the only part of the API where the user's GitHub access token and user ID are exposed for the frontend to use.
This will only return those values for the currently authenticated user.