Use moment to format datetime in publishedState component test
This commit updates the publishedState component's test to use moment to format the date string it expects the component to render.
This is done to make sure the date string is displayed in the current timezone and not in UTC which is what it was hardcoded to do.
With this change the test should pass regardless of the timezone it is run in.