Don't stub `GitHub.registerSite` in integration tests
One of the weaknesses of our integration tests was the fact that they did not cover the GitHub service.
In the integration tests we stubbed the GitHub service's API to prevent it from sending HTTP requests to GitHub which would 404.

This commit fixes that by adding a request mock for the Webhook API endpoint and using that in the specs.

This needs to be used in the factories when one-off sites are created, and in tests where we want to verify that a webhook is created successfully.
It also needs to be used in requests that create Sites.