Add request mock for GET repo GitHub API action
This commit adds a request mock for the GitHub API action that gets a repo.
Currently, we use that request in the API to fetch the permissions for the given user to verify that the user has permission to push to that repo.

Prior to this commit, the integration tests stubbed `GitHub.checkPermissions`.
This commit makes it possible to avoid stubbing this method so that the integration tests can cover the code in that method.