Remove done callback from GitHub.setWebhook
The GitHub.setWebhook callback took a done callback which is not necessary now that it uses a promise.
This commit removes the done callback and handles errors in the promise.
It also changes the code in places where GitHub.setWebhook is called to expect a promise and not a function that handles a callback.