Set nvmrc to have the wrong node version (4) again.
Deploys started failing with this and we need to look further at
it to determine what's going on. Since it's simple to test if the
change we're reverting was actually the cause, get that data
point.

Revert of 8b220c8c1866a9ab3e00f7ef53b89b82c33e1446