Refactor site creation API
This commit refactors the API endpoints that are used to create sites.

Prior to this commit, sites were created using `/v0/site/clone` to create a site from a template, and `/v0/user/add-site` to create a site with an existing repo on GitHub.
This commit changes the API so that both actions can be performed using `/v0/site`.

To create a site from a template, a post request can be sent to `/v0/site` with the following payload:

```json
{
  template: "team",
  owner: siteOwner,
  repository: siteRepository,
  defaultBranch: "master",
  engine: "jekyll"
}
```

The template will be cloned pushed to the new repo.
The site will then be built.

To create a site from an existing GitHub repo, a post request can be sent to `/v0/site` with the following payload:

```json
{
  owner: siteOwner,
  repository: siteRepository,
  defaultBranch: "master",
  engine: "jekyll"
}
```

The site will be cloned from the provided repo and then the site will be built.

Ref #665