Change frontend to use new site creation endpoint
This commit changes the frontend so that it uses the `POST /v0/site`
endpoint to create sites. This meant changing the addSite and cloneRepo
site actions to use the same function on the Federalist API. There were
some other minor changes to get the template name into the API request
body.

ref #665