Fix SiteContainer so that build logs and branches appear correctly
This commit fixes 2 problems involving the site container. These
problems worked together to make the site data as it was displayed in
the container look incorrect.

The first issue was in the site container's code itself. The container
was importing an action creator and not an action to replace the router
history. This meant that if the container loaded without site data
instead of re-directing to the site list as it was intended, it rendered
its children without data fetched from GitHub.

The second issue was deep in the PublishedDate component which was
calling `pop()` on the build array for each site. This was removing the
most recent build from each site so the build logs were rendering
without that build.