Add download-sites-export script
This commit adds a script that uses `cf ssh` to access a running
Federalist instance, execute the `exportSitesAsCsv.js` script, and
download the results as `current-sites.csv`.

This will in effect allow you to see all of the sites on Federalist in
GovCloud by running `./scripts/download-sites-export.sh && cat
current-sites.csv`.