Add build log API
This commit implements the build log API.
This has 2 endpoints, one for creating builds and one for getting a
build's logs.

The create endpoint is meant to be used by build containers after they
run a script.

The find endpoint is mean to be used by the frontend when showing a user
a builds log's.