Add lifecycle callback to filter secrets from build logs
This commit adds a lifecycle callback to the BuildLog that filters any
secerts that may appear in the build logs. We expect the build container
to be built in such a way that secrets appearing in the build logs
should not be a concern. This is a safeguard to protect the logs in the
case where they do.