Refactor preview code with S3Proxy service class
This commit moves the preview logic into a single service class called
S3Proxy. Prior to this commit, this code lived in the preview
controller, previewProxy policy, and sessionAuth policy.

This commit puts all of the logic in the previous places into an S3Proxy
service which does the following:

- Verify that the user is authorized to view the preview
- Resolve the key to use on S3
- Pipe the content from S3 through the response