Handle situation where a build's user is not in site.users array
The build logs where crashing in staging because there was a build with
a user that was not present in the `users` array on the SiteBuild
component's `site` prop. This was causing the following error:

```
Cannot read property "username" of undefined.
```

The source of the error was the `getUsername` method. This commit fixes
that method so that is can handle this exception case, but does not fix
the root issue, that users building sites are not being rendered in the
API response.

ref #695