Add user to a site in response to a webhook
Prior to this commit, when a webhook was received saying for a given
user / site combination, a build was created, but the user was not added
to the site. The result of this was users who had created builds for a
site, but were not associated with the site. This lead to the JSON
response for the site API having a user prop that was missing users that
were referenced by objects in the build prop. This issue became visible
as #695.

This commit fixes the issue by adding a user to a site if they are not
already added in response to a webhook request.