Use consistent API error response format
This commit uses the following response for every API error response:

```json
{
  "status": 403,
  "message": "You are not authorized to do that thing"
}
```

### Error response handlers

This change heavily modifies the response helpers for errors. For generic errors, an error response handler was added which attempts to determine what type of error has occurred and pipe it to the correct response helper. This done by:

1. *Checking the status:* If the error has a status code that matches a supported HTTP error status code, then the error will be piped to the error handler for that status code. Additionally, if the error is simply a number, (e.g. `throw 404`) a generic error for the corresponding status code will be used.
1. *Checking if the error is a Waterline validation error:* Validation errors from Waterline result in a 400.
1. *Creating a generic 500:* If the error response handler cannot make a determination, it renders a 500 with a generic message.

```javascript
// Render a 404
res.error(404)

// Render a 403
res.error({
  status: 403,
  message: "You need admin privileges to do that."
})

// Render a 500
Promise.resolve().then(() => {
  undefined.property
}).catch(err => {
  res.error(err)
})
```

### Controllers

This commit modifies the way errors are handled in controllers to use the new error response handlers.

### Tests

This commit adds the error responses to the Swagger file, which adds a JSON schema for errors. This commit uses that JSON schema to verify that errors match the expected format.