Add AlertBanner to SiteList
This commit changes the SiteList component so that it renders an alert banner when there is an error. This was not the case before which meant errors were not being displayed to users.

This issue was most obvious when there was an error adding a site, since the SiteList component is where a user lands after such an error occurs. Prior to this commit, no error message was visible.

Ref #205