Quit dispatching undefined actions
The `alertActions` and `dispatchActions` were calling `dispatch` with a function that itself called dispatch instead of an action creator. In effect, this is what was happening:

```javascript
dispatch(dispatch(action))
```

Since `dispatch` returns `undefined`, the result was the dispatch of an undefined action. This caused the Redux store to throw an error when it attempted to inspect the action. Specifically, `Cannot read property 'type' of undefined` when it attempted to inspect the action's type.

This commit fixes the problem by removing the outer `dispatch` call.

Ref #713