Disable blueprint
This commit disables blueprint which automatically generates routes and actions in favor of being explicit about routes and actions. The advantages of this are:

1. More clarity about how the app works / what's exposed
1. Better interoperability to another JS web app framework

This commit also removes the `/v0/user/:id` route since it was effectively identical to `/v0/me` and was never used by the frontend.

This commit also adds rules to the JSON schemas to verify that certain properties aren't in response body for users, functionality that was previously wrapped in the tests for `/v0/user/:id`, even though that doesn't make sense.