Use Postgres database in the tests
The tests were running using an in-memory database instead of a postgres database. This was going to create problems when migrating away from Waterline because we'll have to leave behind the in-memory adapter. Additionally, using the in-memory store potentially covers up problems that may occur when we use a SQL based database.

This commit adds a connection to the test configuration to use a postgres database named `federalist-test`. This commit also moves around the testing npm scripts so that the test database is migrated before the tests run.

This commit adds a database chore to drop all records the database is storing before running the tests so that records from previous tests won't cause interference.

The changes in this commit revealed some issues in the test / application logic that were also fixed.