Implement Sequelize model logic and remove Waterline models
This commit adds the logic and tests for the Sequelize models. It also contains a significant rewrite of the factory helper to leverage some of Sequelize's power. Finally, this commit removes all of the old Waterline model code.

Note that this commit is not integrated with the rest of the application. It should not be merged until a changeset that integrates it is added on top.