Update S3Proxy and passport services to use sequelize models
This commit updates the S3Proxy and passport services to use the new
sequelize models instead of the old Waterline ones. Unfortunaltely,
neither of these services have tests at this time, so we are unable to
verify that these changes had the intended effect until the rest of the
application is converted from Waterline to Sequelize.