Apply Sequelize changes to Build Log API
This commit replaces the Waterline code with Sequelize code in order to
get the request specs for the Build Log API green. This involved adding
code to serialize JSON responses for the Build Logs, a pattern which
will need to be repeated in additional commits for the other API
resources.