Apply Sequelize changes to the Site API
This commit updates the Site API such that it uses the new Sequelize
models instead of the old Waterline models. This involved a few changes,
such as moving the call that created the GitHub webhook into the
controller and adding a serializer for site since removing Waterline
also removes its JSON serializers.