Drop merge sites migration
This commit drops the merge sites migration. This migration does not modify the schema, but rather adds / deletes records. This migration will not work moving forward since it depends on Waterline code.

Since all deployed versions of Federalist that we know about have run this migration, and all development version should be able to drop / recreate their database without issue, it should be safe to drop this migration.