Cleanup dead code after deleting the editor
This commit pulls out all of the dead code that was orphaned when the editor was disabled. This commit also updates the tests so they don't fail without the removed files.

I tested this pretty thoroughly, but we may want to watch on staging for a while before merging into master.