Show correct error message when a user does not have GitHub admin access
Prior to this commit, when a user did not have admin access to a
repository, the user would see an error with the message "Internal
Server Error". The reason for this was two-fold:

1. Waterline wrapped errors that occured in lifecycle callback in a WL_ERROR.
2. The error was not setting its status code so it was interpretted as a 500

With Waterline gone, 1 is no longer a problem. This commit fixes 2.

Additionally, this commit adds a test for this case.

ref #743