Extract sails-db-migrate
This commit removes the sails-db-migrate node module, and replaces it with a `migrate.js` file which does the same thing, except it does not load Sails and it more specific to our needs.

This commit also strips out Grunt which as far as I can tell was only being used to run migrations.