Extract Sails configuration management
This commit begins the process of moving our custom configuration out of Sails. This is done by moving configs that Sails does not need to lift out of the `config` directory and into a separate `configg` directory. The `configg` directory has an `index.js` file which loads the configs in a similar way to Sails.

Ref #194