Extract Sails and setup Express middleware stack
This commit removes Sails and a lot of the cruft that was growing alongside it. The new middleware stack includes middleware that does essentially what Sails did for us, plus a custom middleware to give us our response helpers and a custom middleware to give us routing.

The routing middleware is currently in a TODO state.

Ref #763