Fix error message for non-existant repo
When a repo didn't exist, the API would render an error message saying
"The repo <org name>/null does not exist`. This was because the null
repo object and the repo passed into the function checking the repos
existence had a name collision. Because of this, the null repo object
was being put in the error message.

This commit fixes the issue by using a different name for the repo
object fetched from github.

Ref #765