Send build log probs to SiteBuildLogs component
The build logs where not being set in the props of the SiteBuildLogs
container. This happened b/c the SiteContainer was not able to determine
which component it was rendering because the `getPageTitle` function was
not returning.

This commit fixes the above by adding a `return` to the `getPageTitle`
function.