Reduce instance memory from 512MB to 256MB
The Federalist web application uses about 150MB of RAM. With this in mind, this commit scales the memory the app is allocated in cloud.gov from 512MB to 256MB, which should free up some memory in our space, while leaving overhead from memory spikes.

This commit also bumps the instance count in production to 5.