Recognize empty strings in update site request body
The Site API was not updating attributes for a site if the new value was an empty string. This was because values where being set as such:

```javascript
{
  config: req.body.config || site.config,
}
```

Since the empty string evaluates to false, this was falling back to the value on the site. This commit fixes the problem by creating a params object that is the result of merging the site and request body:

```javascript
const params = Object.assign(site, req.body)
{
  config: params.config,
}
```

Ref #759