Filter incomplete builds before sorting to find last published date
The publishedState component was displaying "Please wait for build to complete or check logs for error message." for sites with builds that had been published. This component is intended to show the `completedAt` date for the last published build.

The issue was that these sites had builds that had yet to complete, meaning their `completedAt` prop was `undefined`. When the builds were being sorted to get the most recently completed build, the builds with `undefined` for their completed at date always went at the end, meaning they were recognized as the most recent complete build. This commit fixes the issue by filtering out builds with an `undefined` completed at date before sorting builds.

FWIW, the reason the sort function was broken is that in JavaScript `1 < undefined` and `1 > undefined` both return false somehow.

Ref #793