Use correct default branch for template sites
Not all of the templates use `master` as their default branch, but prior to this commit, we used `master` as the default branch for every site created from a template. This commit adjusts the SiteCreator so that is looks up a template and uses the branch specified by the template as the default branch.

Ref #799