Add published branch API
This commit adds an API endpoint that provides the client with a list of the branches that are currently published to S3. This includes a list of published preivew branches and the default branch.

The list of preview branches is generated by asking the S3 API for a list of folders in the site's preview directory on S3. This is done using a new S3PublishedFilesLister service.