Respect path component of custom domains
Prior to this commit, Federalist did not respect the path component of custom URLs. This caused problems for sites that were not displayed at the root for their domain, e.g. `www.opm.gov/opm-site`. This commit modifies the service that sends builds to the build scheduler to send the path component for custom domains.

ref #812