Add ability for published branch serializer to serialize a single branch
Prior to this commit, the serializer for published branches only handled an array of branch names and had no concept of files for a given branch. This commit changes it so that it can serialize a single branch and include a list of files on that branch as part of its result.