Add findOne action to published branches API
This commit adds an action to the published branches API that can be used to fetch data about a specific branch that has been published to S3. The response contains all of the data contained in the response to the `find` action, with the addition of an array of files that have been published to S3 for the given branch.