Redirect to preview after sign in
Prior to this commit, users who were not signed in attempting to look at a preview were redirected to the home page. In order to see the preview, they would have to login, then go back to see the preview.

This commit smooths that process out by redirecting back to previews after sign in. It does this by saving the URL the user was attempting to visit in the session, then redirecting to the sign in path. After sign in, it redirects back to the URL.

Ref #117