Show correct error message when user attempts to add a site
Prior to this commit, Federalist would show an error message that read "You do not have write access to the repository" when a user attempted to add a new site. This suggested that the user only need attain write access to add a site, when in reality the first user to add a site needs Admin access.

This commit changes the logic used to display that error message to behave as follows:

- If the site is brand new, and the user does not have admin access, show an admin access error message
- If the site already exists, and the user does not have write access, show a write access error message
- If the repo does not exist, show a nonexistant repo error message

Ref #719