Use Federalist preview URLs instead of S3 URLs for view links in branch responses
The view links in the API responses for previews were S3 URLs, which don't support the public/private preview functionality. This commit adds an env variables named `APP_HOSTNAME` which tells Federalsit what its hostname is. It then uses that hostname to generate a Federalist preview URL instead of an S3 one.

In a future commit, we should consider changing the build callback URL, build token URL, and GitHub callback URL to use the `APP_HOSTNAME` variable instead of needing to be set in the environment.

Ref #825