Allow sites to specify an image path for template images
This commit adds a `thumb` prop to the template configuration. Here, a link to a thumbnail image for the template can be set. This will allow template maintainers to link to external template URLs and avoid having to PR Federalist to change the thumbnail when the template changes.

This commit should be followed up with commits to link to template thumbs in the templates' respective repos so they can be maintained there.