Don't bind to localhost:1337 during tests
This commit changes the request specs so they require the app and pass it into supertest instead of listening on a port and pointing supertest at that. This has the advantage of making the app more portable and means the test suite can run at the same time as the app.