Lowercase owner and repo before looking for existing site at site creation
The check to make sure a site was not added twice by looking for the owner / repo name was broken if the name of the site to be added contained uppercase characters. This was because we lowercase the name of the site / repo when adding then, but not when checking for their existence.

This commit lowercases the name of the owner and repo before checking for their existence.

ref #851